interim; 'firebase-prime' CLI tool moved, builds into a Dockerfile
