+++ firebase-ci-builder.user/README.md
@@ -1,0 +1,10 @@
+# `firebase-js-builder.user`
+
+Docker image for *using* the `firebase-ci-builder`.
+
+Gets automatically pulled and built, by `docker compose up`, in various subpackages.
+
+
+
+
+
