+++ firebase-ci-builder.user/firebase-prime/.gitignore
@@ -1,0 +1,2 @@
+/out/
+
