--- packages/app/tools/firebase-prime/.npmrc
@@ -1,1 +1,0 @@
-package-lock=false
