--- packages/app/tools/firebase-prime/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "firebase-prime",
-  "description": "Priming an emulated Firestore instance with users and data, from ESM files.",
-  "private": true,
-  "type": "module",
-  "bin": {
-    "firebase-prime": "./index.js"
-  },
-  "dependencies": {
-    "commander": "^7.2.0"
-  },
-  "//": {
-    "usage": [
-      "'prime --project=xyz docs.js users.js"
-    ]
-  }
-}
